第一次尝试提交
但是由于不能用 Xcode Beta 来提交，所以有问题啊。

+++ 诗词歌赋/PoetModel.swift
@@ -16,7 +16,7 @@ class PoetModel : NSManagedObject {
     @NSManaged var title:String
     @NSManaged var context:String
 	
-    @NSManaged var score:Int // User's rate score: range [-1, 100]
+    @NSManaged var score:Int64 // User's rate score: range [-1, 100]
 	// -1 为讨厌。 0 为默认值。 分数越高越喜欢
